requirements/common.txt
Django==1.11.1
requests==2.14.2
python-dateutil==2.6.0
